quick-service-sash/services/backend/bot_chat/requirements.txt
Flask==3.1.0
ipython==8.12.3
langchain_core==0.3.30
langchain_openai==0.3.1
langgraph==0.2.64
python-dotenv==1.0.1
typing_extensions==4.12.2
langgraph-checkpoint-postgres==2.0.9
psycopg[binary,pool]